requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
librosa==0.10.1
opencv-python-headless==4.8.1.78
deepface==0.0.79
websockets==11.0.2
python-multipart==0.0.6
python-dotenv==1.0.0
joblib==1.3.2
